requirements.txt
gunicorn==20.1.0
label-studio-ml @ git+https://github.com/HumanSignal/label-studio-ml-backend.git
ultralytics
numpy
pyyaml
pillow
